requirements.txt
pandas == 1.1.2
Pillow == 7.2.0
pandastable==0.12.2.post1
matplotlib==3.1.1
opencv-python==4.4.0.44
Keras== 2.4.3
tensorflow==2.3.0
numpy==1.18.5

